Fixing bug #1
With the old implementation other developers using the
member model had a problem, because it still contained
the cached values.

Now i am using the member model itself to write the
new value back to the database. So all other extensions
should have the value in the member model as well.

Signed-off-by: Leo Unglaub <leo.unglaub@2create.at>